Abandoned Cars in Chicago

Information about abandoned vehicles in Chicago, Illinois. Includes all open abandoned vehicle service complaints made to 311 since January 2011, and all requests completed since January 1 2011.

abandonedcars

Format

A data frame with 137867 rows and 21 variables:

Creation Date

Date that the complaint was made

Status

Status of the request. Open or complete.

Completion Date

Date the request was completed. Left blank if not completed.

Service Request Number

Unique identifier for each service request.

Type of Service Request

Abandoned Vehicle Complaint for all entries.

License Plate

License plate number and sometimes a description (state, color). Some entries did not include a plate number.

Vehicle Make/Model

The make of the vehicle. Many entries only included the make, not the model.

Vehicle Color

Color of the vehicle

Current Activity

Either place seven day sticker or FVI - Outcome

Most Recent Action

The most recent action taken on this vehicle

How Many Days Has The Vehicle Been Reported as Parked?

Number of days the vehicle has been reported as parked

Street Address

Street Address where the vehicle was found

ZIP Code

Five digit zip code corresponding to the address

X Coordinate

X coordinate of the address

Y Coordinate

Y coordinate of the address

Ward

Aldermanic ward (1-50) corresponding to the address

Police District

Police district (1-25) corresponding to the address

Community Area

Community area (1-77) corresponding to the address

Latitude

Latitude corresponding to address

Longitude

Longitude corresponding to address

Location

Coordinates of the location in format of (Latitude, Longitude)

Source

Retrieved from https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Community-Areas-current-/cauq-8yn6

Details

Dataframe with longitude and latitude variables.