Malaria in Colombia (1998)

Malaria incidence in 1998, population in Colombia from 1973, 1995, 1993 censuses, and projected population from 1995 to 2005. Use list2env(malariaCol, .GlobalEnv) to unlist in working environment. The data consist of a list of two objects:

malariaDept A shapefile object of class sf Dataframe with Colombian departments.

malariaMunic A shapefile object of class sf Dataframe with Colombian municipalities.

malariaCol

Format

malariaMunic includes 1,068 rows, 50 variables, and a geometry column. malariaDept includes 33 rows, 46 variables, and a geometry column:

ID

Sequencial ID

ADM0

Country name

ADM1

Departmentos=States (Administrative level 1)

ADM2

Municipioss=Counties (Administrative level 2)

IDDANE

DANE (Colombian National Bureau of Statistics) code for counties in numeric format

DANECODE

DANE (Colombian National Bureau of Statistics) code for counties in character format

CODDEPT

DANE (Colombian National Bureau of Statistics) code for states

TPyyyy

Total population from census in year 'yyyy' (1973, 1985, and 1993 have census population, the remaning years have projected population)

UPyyyy

Urban population from census in year 'yyyy' (1973, 1985, and 1993 have census population, the remaning years have projected population)

RPyyyy

Rural population from census in year 'yyyy' (1973, 1985, and 1993 have census population, the remaning years have projected population)

MALARI98

Malaria incidence in 1998

Source

SIVIGILA. Ministerio de Salud, Colombia. There are 3 files for each year from 1995-2005. Projections from Departamento Administrativo Nacional de Estadistica DANE, Colombia https://www.minsalud.gov.co and https://www.dane.gov.co/index.php/estadisticas-por-tema/demografia-y-poblacion/proyecciones-de-poblacion

Details

List of two Sf objects, unprojected. EPSG 4326: WGS84.

Examples