Rental Housing and Demographics in NYC (2000s) (non-spatial)

Demographic and housing data for New York City’s 55 sub-boroughs (2000s).

nyc

Format

A data frame with 55 rows and 34 variables:

CODE

sub-borough code, 1XX Bronx, 2XX Brooklyn, 3XX Manhattan, 4XX Queens, 5XX Staten Island

FORHIS06

percentage of hispanic population, not born in US, 2006

FORHIS07

percentage of hispanic population, not born in US, 2007

FORHIS08

percentage of hispanic population, not born in US, 2008

FORHIS09

percentage of hispanic population, not born in US, 2009

FORWH06

percentage of white population, not born in US, 2006

FORWH07

percentage of white population, not born in US, 2007

FORWH08

percentage of white population, not born in US, 2008

FORWH09

percentage of white population, not born in US, 2009

HHSIZ1990

average number of people per household, 1990

HHSIZ00

average number of people per household, 2000

HHSIZ02

average number of people per household, 2002

HHSIZ05

average number of people per household, 2005

HHSIZ08

average number of people per household, 2008

KIDS2000

percentage households w kids under 18, 2000

KIDS2005

percentage households w kids under 18, 2005

KIDS2006

percentage households w kids under 18, 2006

KIDS2007

percentage households w kids under 18, 2007

KIDS2008

percentage households w kids under 18, 2008

KIDS2009

percentage households w kids under 18, 2009

NAME

name of borough, one of five

RENT2002

median monthly contract rent, 2002

RENT2005

median monthly contract rent, 2005

RENT2008

median monthly contract rent, 2008

RENTPCT02

percentage of housing stock that is market rate rental units, 2002

RENTPCT05

percentage of housing stock that is market rate rental units, 2005

RENTPCT08

percentage of housing stock that is market rate rental units, 2008

SUBBOROUGH

name of sub-borough

PUBAST90

percentage of households receiving public assistance, 1990

PUBAST00

percentage of households receiving public assistance, 2000

YRHOM02

average number of years living in current residence, 2002

YRHOM05

average number of years living in current residence, 2005

YRHOM08

average number of years living in current residence, 2008

bor_subb

sub-borough code, repeated

Source

https://geodacenter.github.io/data-and-lab/nyc/

Details

Dataframe, no spatial components.