John Snow & the 19th Century Cholera Epidemic

Datasets about the Broad St Pump and South London Natural Experiment cases in London (Snow, 1855). Use list2env(snow, .GlobalEnv) to unlist in working environment. The data consist of a list of six objects:

snow2 A shapefile object of class sf Dataframe with spatial points of buildings where cholera deaths were recorded near the Broad Street pump.

snow3 A shapefile object of class sf Dataframe with polygons of housing blocks, in the forms of polygons, which aggregate cholera deaths in the vicinity of the Broad Street pump.

snow4 A shapefile object of class sf Dataframe with spatial points that represent the Broad Street pump and the five pumps closest to it.

snow5 A shapefile object of class sf Dataframe with rings of individual cholera deaths and the Broad Street pump, extracted from snow4.

snow6 A shapefile object of class sf Dataframe with rings of individual cholera deaths and the five pumps except for the Broad Street pump, extracted from the 6 pumps dataset (from snow4).

snow7 A shapefile object of class sf Dataframe with polygons that correspond to the South London Natural Experiment attribute data compiled by Coleman (2019).

snow

Format

snow2 includes 250 rows, 8 variables, and a geometry column:

ID

Unique ID

x

X coordinates (in meters)

y

Y coordinates (in meters)

deaths

Number of deaths per building

pumpID

ID of the nearest pump (see ‘pumps’ dataset)

distpump

Distance to the nearest pump (in meters - see ‘pumps’ dataset)

distBSpump

Distance to Broad St pump (in meters - see ‘pumps’ dataset)

BSpump

Create categories depending on whether the Broad Street pump is closest (1) or not (0)

geometry

Point

snow3 includes 40 rows, 3 variables, and a geometry column:

ID

Unique ID

deaths

Number of deaths per polygon

deathdens

Number of deaths per polygon divided by population

geometry

Polygon

snow4 includes 6 rows, 4 variables, and a geometry column:

ID

Pump ID

x

X coordinates (in meters)

y

Y coordinates (in meters)

name

Name of the pump

geometry

Point

snow5 includes 60 rows, 4 variables, and a geometry column:

ID

Unique ID

area

Area (in squared meters)

deaths

Number of deaths per building

deathdens

Number of deaths per ring divided by area

geometry

Polygon

snow6 includes 35 rows, 6 variables, and a geometry column:

ID

Pump ID

x

X coordinates (in meters)

y

Y coordinates (in meters)

dist

Distance to pump (in meters)

deaths

Number of deaths per ring

geometry

Polygon

snow7 includes 32 rows, 32 variables, and a geometry column:

dis_ID

London district ID

district

London district

sub_ID

London subdistrict ID

subdist

London subdistrict

pop1851

Population for 1851

supplier

Water company suppliers that served the subdistrict

supplierID

Water company supplier ID

perc_sou

Proportion of the population that was served by the Southwark & Vauxhall company

perc_lam

Proportion of the population that was served by the Lambeth company

perc_other

Proportion of the population that was served by a company other than Southwark & Vauxhall or Lambeth

lam_degree

Creates categories for the proportion of the population that was served by the Lambeth company

d_overall

Number of deaths attributed to the cholera epidemic in the seven weeks ending August 26, 1854

d_sou

Number of deaths attributed to the cholera epidemic in the seven weeks ending August 26, 1854 for the Southwark & Vauxhall company

d_lam

Number of deaths attributed to the cholera epidemic in the seven weeks ending August 26, 1854 for the Lambeth company

d_pump

Number of deaths attributed to the cholera epidemic in the seven weeks ending August 26, 1854 originating in pump-wells

d_thames

Number of deaths attributed to the cholera epidemic in the seven weeks ending August 26, 1854 from water from the Thames River and ditches

rate_sou7w

Southwark & Vauxhall cholera death rate per 10000 people in the seven weeks ending August 26, 1854

rate_lam7w

Lambeth cholera death rate per 10000 people in the seven weeks ending August 26, 1854 - Missing values are undefined and should not be converted to 0

rate_oth7w

Cholera death rate per 10000 people for ‘other’ category in the seven weeks ending August 26, 1854 - Missing values are undefined and should not be converted to 0

deaths1849

Number of deaths attributed to the cholera epidemic in 1849

deaths1854

Number of deaths attributed to the cholera epidemic in 1854

rate1849

Cholera death rate per 10000 people in 1849

rate1854

Cholera death rate per 10000 people in 1854

pop1849

Population for 1849

pop1854

Population for 1854

rAvSupR_49

Average supplier-region-specific cholera mortality rate per 10000 people in 1849

rAvSupR_54

Average supplier-region-specific cholera mortality rate per 10000 people in 1854

pred_Snow

Snow’s cholera death count prediction (from his 1856 Table VI)

pred_DiD49

Cholera death count prediction from Difference-in-Difference regression analysis for 1849

pred_DiD54

Cholera death count prediction from Difference-in-Difference regression analysis for 1854

geometry

Polygon

Source

A detailed description of the datasets and their construction can be found at: https://geodacenter.github.io/data-and-lab/snow. The GeoData center also has available snow1, which contains 578 individual deaths during the 1854 cholera epidemic that were compiled by Tobler (1994). snow1 was excluded from this package given that it can be reproduced from R package HistData, version 0.8-6.

Details

List of six Sf objects, projected to OSGB 1936 / British National Grid.

Examples